groq_qna_rag/requirements.txt
langchain_groq
langchain
python-dotenv
langchain_community
langchain_core
streamlit
pypdf
langchain_openai
faiss-cpu